// src/app/models/act.model.ts
import {ActTitle} from "./ActTitle";

export interface Act {
  id?: number;            // Optional because when creating a new act, the ID will be generated by the backend
  title: ActTitle;        // The title is an object of type `ActTitle`
  description: string;    // Description specific to the appointment
  appointmentId: number;  // The ID of the associated appointment
}